refactoring: extract IvsPlayerVideo component
