+++ src/components/IvsPlayer/__tests__/IvsPlayer.spec.tsx
@@ -8,7 +8,6 @@ test('IvsPlayer', () => {
       playBackUrl={
         'https://2a78e417f923.us-west-2.playback.live-video.net/api/video/v1/us-west-2.607167088920.channel.QFdTdOPixOx2.m3u8'
       }
-      autoplay={true}
     />,
   )
   const tree = component.toJSON()
